feat: move files to correct sub dir
