+++ libs/coin-modules/coin-xrp/src/signer/getAddress.ts
@@ -1,9 +1,9 @@
 import { GetAddressOptions } from "@ledgerhq/coin-framework/derivation";
 import { GetAddressFn } from "@ledgerhq/coin-framework/bridge/getAddressWrapper";
 import { SignerContext } from "@ledgerhq/coin-framework/signer";
-import { XrpAddress, XrpSigner } from "./signer";
+import { XrpAddress, XrpSigner } from "../types/signer";
 
-const resolver = (signerContext: SignerContext<XrpSigner>): GetAddressFn => {
+const getAddress = (signerContext: SignerContext<XrpSigner>): GetAddressFn => {
   return async (deviceId: string, { path, verify }: GetAddressOptions) => {
     const { address, publicKey } = (await signerContext(deviceId, signer =>
       signer.getAddress(path, verify, false),
@@ -17,4 +17,4 @@ const resolver = (signerContext: SignerContext<XrpSigner>): GetAddressFn => {
   };
 };
 
-export default resolver;
+export default getAddress;
