+++ libs/coin-modules/coin-xrp/src/signer/index.ts
@@ -1,0 +1,7 @@
+/**
+ * This directory is the home for all types and logic based on Ledgers signer.
+ */
+
+import getAddress from "./getAddress";
+
+export default getAddress;
