+++ libs/coin-modules/coin-xrp/src/test/bot-specs.ts
