+++ libs/coin-modules/coin-xrp/src/test/cli.ts
