+++ libs/coin-modules/coin-xrp/src/types/bridge.ts
