+++ libs/coin-modules/coin-xrp/src/types/index.ts
@@ -1,0 +1,2 @@
+export * from "./bridge";
+export * from "./signer";
