+++ libs/coin-modules/coin-xrp/src/types/signer.ts
